@@ -15,6 +15,6 @@
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
-    "electron": "^1.8.1"
+    "electron": "^5.0.1"
   }
 }
