@@ -1,25 +1,21 @@
-/* jslint esversion:6, node: true */
-/* jshint -W117 */
-'use strict';
+"use strict";
 
-const {
-  ipcRenderer
-} = require('electron');
+import { ipcRenderer } from "electron";
 
-var btn = document.querySelector('button.new-game');
-btn.addEventListener('click', function (e) {
+var btn = document.querySelector("button.new-game");
+btn.addEventListener("click", function(e) {
   e.preventDefault();
   e.stopPropagation();
-  ipcRenderer.send('game-request', 'new-game');
+  ipcRenderer.send("game-request", "new-game");
 });
 
 function makeGameCell(data, idx) {
-  let cell = document.createElement('div');
+  let cell = document.createElement("div");
   if (data === null) {
-    cell.className = 'cell-data blank-cell';
-    cell.appendChild(document.createTextNode(''));
+    cell.className = "cell-data blank-cell";
+    cell.appendChild(document.createTextNode(""));
   } else {
-    cell.className = 'cell-data';
+    cell.className = "cell-data";
     cell.id = `cell-${idx}`;
     cell.appendChild(document.createTextNode(data));
   }
@@ -34,17 +30,17 @@ function makeGameCard(parent, words) {
       if (row !== undefined) {
         parent.appendChild(row);
       }
-      row = document.createElement('div');
-      row.className = 'row';
+      row = document.createElement("div");
+      row.className = "row";
     }
     row.appendChild(makeGameCell(words[i], i));
   }
   parent.appendChild(row);
 }
 
-ipcRenderer.on('game-data', function (evt, msg) {
+ipcRenderer.on("game-data", function(evt, msg) {
   var gameWords = JSON.parse(msg);
-  var gameParent = document.querySelector('div.game-card');
+  var gameParent = document.querySelector("div.game-card");
   gameParent.innerHTML = null;
   makeGameCard(gameParent, gameWords);
 });
@@ -58,16 +54,16 @@ function getCell(evt) {
 }
 
 function toggleSeen(cell) {
-  if (cell.className.includes('seen-word')) {
-    cell.className = 'cell-data';
+  if (cell.className.includes("seen-word")) {
+    cell.className = "cell-data";
   } else {
-    cell.className = 'cell-data seen-word';
+    cell.className = "cell-data seen-word";
   }
 }
 
 // deal with completed words
-var gameDiv = document.querySelector('div.game-card');
-gameDiv.addEventListener('click', function (evt) {
+var gameDiv = document.querySelector("div.game-card");
+gameDiv.addEventListener("click", function(evt) {
   var cellElement;
 
   evt.preventDefault();
